update files order services

+++ src/main/java/com/jersson/arrivasplata/swtvap/api/order/business/implementation/OrderDetailServiceImpl.java
@@ -1,11 +1,15 @@
 package com.jersson.arrivasplata.swtvap.api.order.business.implementation;
 
 import com.jersson.arrivasplata.swtvap.api.order.business.service.OrderDetailService;
+import com.jersson.arrivasplata.swtvap.api.order.exception.CustomException;
 import com.jersson.arrivasplata.swtvap.api.order.model.OrderDetail;
 import com.jersson.arrivasplata.swtvap.api.order.repository.OrderDetailRepository;
+import org.springframework.stereotype.Service;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 
+import java.util.Optional;
+@Service
 public class OrderDetailServiceImpl implements OrderDetailService {
 
     private final OrderDetailRepository orderDetailRepository;
@@ -21,16 +25,32 @@ public class OrderDetailServiceImpl implements OrderDetailService {
 
     @Override
     public Mono<OrderDetail> findById(Long id) {
-        return Mono.justOrEmpty(orderDetailRepository.findById(id));
+        return Mono.just(orderDetailRepository.findById(id)
+                .orElseThrow(() -> new CustomException("OrderDetail not found with id: " + id)));
     }
 
     @Override
-    public Mono<OrderDetail> save(OrderDetail orderDetail) {
-        return Mono.justOrEmpty(orderDetailRepository.save(orderDetail));
+    public Mono<OrderDetail> save(OrderDetail orderDetail) { //Guardar
+        if (orderDetail.getProductId() == null || orderDetail.getOrderId() == null ) {
+            throw new CustomException("OrderDetail ProductId || OrderId is required");
+        }
+        return Mono.just(orderDetailRepository.save(orderDetail));
+    }
+
+    @Override
+    public Mono<OrderDetail> updateOrderDetail(OrderDetail orderDetail) {
+        if (orderDetail.getOrderDetailId() == null ) {
+            throw new CustomException("OrderDetail id is required");
+        }
+        return Mono.just(orderDetailRepository.save(orderDetail));
     }
 
     @Override
     public Mono<Void> deleteById(Long id) {
+        Optional<OrderDetail> orderDetail = orderDetailRepository.findById(id);
+        if (!orderDetail.isPresent()) {
+            throw new CustomException("OrderDetail not found with id: " + id);
+        }
         orderDetailRepository.deleteById(id);
         return Mono.empty();
     }
